requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Retrieval
sentence-transformers==2.7.0
faiss-cpu==1.8.0
rank-bm25==0.2.2

# Generation
transformers==4.40.0
accelerate==0.30.1
torch==2.2.2

# Evaluation
ragas==0.1.7
datasets==2.19.0
openai==1.25.0

# Interpretability
matplotlib==3.8.4
numpy==1.26.4

# Document loading
pypdf==4.2.0

# UI
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
